% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_col_breakdown_plot.R
\name{avg_col_breakdown_plot}
\alias{avg_col_breakdown_plot}
\title{Function that plots the average living costs per state}
\usage{
avg_col_breakdown_plot(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A pie chart of the average living costs for a single person
earning minimum wage for one state
}
\description{
A function that plots the average cost of living for a single person  at one state,
given a state as input.
}
\examples{
avg_col_breakdown_plot("MA")

}
