% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_functions.R
\name{avg_col_multiple_states}
\alias{avg_col_multiple_states}
\title{Function to calculate gross salary, tax, net salary, and difference between average cost of living
and net salary across multiple states}
\usage{
avg_col_multiple_states(states)
}
\arguments{
\item{states}{A list of states}
}
\value{
A tibble of minimum wage, gross salaries earning minimum wage, federal,
state, local taxes, net salaries, and difference between average cost of living
and net salary per state
}
\description{
A function that calculates the annual gross salary, total tax (federal, state, local), net salary,
and difference between average cost of living and net salary
of a single person earning minimum wage at one or more states, given a vector of states as input.
}
\examples{
avg_col_multiple_states(c("MA","TX","IL","AL"))
avg_col_multiple_states("MA")

}
