% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_col_multiple_states_plot.R
\name{avg_col_multiple_states_plot}
\alias{avg_col_multiple_states_plot}
\title{Function that plots the difference in average cost of living and net salary}
\usage{
avg_col_multiple_states_plot(states)
}
\arguments{
\item{states}{A list of states}
}
\value{
A scatter plot of the difference in average cost of living
for one or multiple states
}
\description{
A function that plots the difference in average cost of living and net salary
of a single person earning minimum wage at one or more states, given a vector of states as input.
}
\examples{
avg_col_multiple_states_plot(c("MA","TX","IL","AL"))

}
