% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_functions.R
\name{federal_tax_owed}
\alias{federal_tax_owed}
\title{Function to calculate tax owed by an individual based on their gross income}
\usage{
federal_tax_owed(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A tibble of taxes, average living costs, minimum wage, and gross
salaries earning minimum wage per state
}
\description{
A function that calculates the amount of federal tax owed by an individual earning
minimum wage at a specific state, given a state as input.
}
\examples{
federal_tax_owed("TX")
federal_tax_owed("MA")

}
