% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salaryFunctions.R
\name{grossSalary}
\alias{grossSalary}
\title{A function that calculates the annual gross salary of a single person earning
minimum wage at a specific state, given a state as input.}
\usage{
grossSalary(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A tibble of taxes, average living costs, minimum wage, and gross
salaries earning minimum wage per state
}
\description{
grossSalary is an added column to the existing us_cost_of_living data set.
This serves as a helper function to the following grossSalaryMultiple()
function.
}
\examples{
grossSalary("MA")
grossSalary("TX")

}
