% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salaryFunctions.R
\name{grossSalaryMultiple}
\alias{grossSalaryMultiple}
\title{A function that calculates the annual gross salary of a single person earning
minimum wage at one or more states, given a vector of states as input.}
\usage{
grossSalaryMultiple(states)
}
\arguments{
\item{states}{A list of states}
}
\value{
A tibble of taxes, average living costs, minimum wage, and gross
salaries earning minimum wage per state
}
\description{
A function that calculates the annual gross salary of a single person earning
minimum wage at one or more states, given a vector of states as input.
}
\examples{
grossSalaryMultiple(c("MA","TX","IL","AL"))
grossSalaryMultiple("MA")

}
