% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_functions.R
\name{net_salary}
\alias{net_salary}
\title{Function to calculate the total tax owed and net salary of an individual based on their gross income}
\usage{
net_salary(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A tibble of the state, minimum wage, gross salary,
total tax owed based on income, and net salary
}
\description{
A function that calculates the combined amount of federal, state, and local tax owed by an individual earning
minimum wage, and their net salary, given a state as input.
}
\examples{
net_salary("TX")
net_salary("MA")

}
