% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_functions.R
\name{net_salary_avg_col_difference}
\alias{net_salary_avg_col_difference}
\title{Function to calculate the difference between average cost of living and net salary}
\usage{
net_salary_avg_col_difference(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A tibble of the state, minimum wage, gross salary,
minimum wage, total tax, net salary,
and the difference between net salary and average cost of living
}
\description{
A function that calculates the difference between the average costs of living (grocery, medical, housing, etc)
by net salary, given a state as input.
}
\examples{
net_salary_avg_col_difference("TX")
net_salary_avg_col_difference("MA")

}
