% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{taxable_income}
\alias{taxable_income}
\title{US tax brackets and federal income tax rates for a single filer}
\format{
\subsection{\code{taxable_income}}{

A data frame with 7 rows representing 7 tax brackets and 4 columns
\describe{
\item{tax_rate}{Federal tax rate}
\item{min_income}{Minimum income threshold for a tax rate}
\item{max_income}{Maximum income threshold for a tax rate}
\item{tax_info}{Additional information regarding the tax owed for each bracket}
}
}
}
\source{
Download from
\url{https://www.nerdwallet.com/article/taxes/federal-income-tax-brackets},
}
\usage{
taxable_income
}
\description{
A data set that contains information about tax brackets based on income and federal income tax rates
}
\keyword{datasets}
