% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{us_cost_of_living}
\alias{us_cost_of_living}
\title{US cost of living data set}
\format{
\subsection{\code{us_cost_of_living}}{

A data frame with 50 rows representing 50 states and 9 columns
\describe{
\item{state}{State name}
\item{state_abbr}{State name abbreviation}
\item{grocery_cost}{State's average grocery cost}
\item{medical_cost}{State's average medical cost}
\item{housing_cost}{State's average housing cost}
\item{transportation_cost}{State's average transportation cost}
\item{civic_cost}{State's average civic cost}
\item{other_cost}{State's average miscellaneous/other cost}
\item{total_avg_cost}{State's total average cost of living}
}
}
}
\source{
Download from
\url{https://livingwage.mit.edu/}
}
\usage{
us_cost_of_living
}
\description{
A data set that contains information related to cost of living in the US,
broken down into different categories.
}
\keyword{datasets}
