% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{us_minimum_wage}
\alias{us_minimum_wage}
\title{US minimum wage data set}
\format{
\subsection{\code{us_minimum_wage}}{

A data frame with 50 rows representing 50 states and 3 columns
\describe{
\item{state}{State name}
\item{state_abbr}{State name abbreviation}
\item{minimum_wage}{State's minimum wage (hourly rate)}
}
}
}
\source{
Download from
\url{https://www.laborlawcenter.com/state-minimum-wage-rates}.
}
\usage{
us_minimum_wage
}
\description{
A data set that contains information related to minimum wage rates in the US.
}
\keyword{datasets}
