% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{us_state_local_tax_rate}
\alias{us_state_local_tax_rate}
\title{US state and average local tax rate data set}
\format{
\subsection{\code{us_state_local_tax_rate}}{

A data frame with 50 rows representing 50 states and 5 columns
\describe{
\item{state}{State name}
\item{state_abbr}{State name abbreviation}
\item{state_tax_rate}{State's tax rate}
\item{avg_local_tax_rate}{State's average local tax rate}
\item{combined_tax_rate}{State's combined tax rate}
}
}
}
\source{
Download from
\url{https://worldpopulationreview.com/state-rankings/cost-of-living-index-by-state},
\url{https://worldpopulationreview.com/state-rankings/cost-of-living-index-by-state}, and
\url{https://www.laborlawcenter.com/state-minimum-wage-rates}.
}
\usage{
us_state_local_tax_rate
}
\description{
A data set that contains information related to US state and average local tax rate data set.
}
\keyword{datasets}
