% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/salary_functions.R
\name{federal_tax_owed}
\alias{federal_tax_owed}
\title{Function to calculate the federal tax owed by an individual based on their gross income}
\usage{
federal_tax_owed(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A tibble of the state, minimum wage, gross salary calculation
earning minimum wage, and the federal tax owed based on income
}
\description{
A function that calculates the amount of federal tax owed by an individual earning
minimum wage based on federal income tax brackets, given a state as input.
}
\examples{
federal_tax_owed("TX")
federal_tax_owed("MA")

}
