% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gross_salary.R
\name{gross_salary}
\alias{gross_salary}
\title{Function to calculate gross salary of one state}
\usage{
gross_salary(input_state)
}
\arguments{
\item{input_state}{A state name, abbreviated (e.g. MA)}
}
\value{
A tibble of the state, minimum wage, and gross salary calculation
earning minimum wage
}
\description{
A function that calculates the annual gross salary of a single person earning
minimum wage at a specific state, given a state as input.
\code{gross_salary} is an added column to the existing \code{us_minimum_wage} data set.
This serves as a helper function to the following \code{avg_col_multiple_states()}
function.
}
\examples{
gross_salary("MA")
gross_salary("TX")

}
