% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{us_cost_of_living}
\alias{us_cost_of_living}
\title{US cost of living data set}
\format{
\subsection{\code{us_cost_of_living}}{

A data frame with 50 rows representing 50 states and 9 columns
\describe{
\item{state}{State name}
\item{state_abbr}{State name abbreviation}
\item{cost_index}{State's cost of living index, with the baseline set at 100}
\item{grocery_cost_index}{State's grocery cost index, with the baseline set at 100}
\item{housing_cost_index}{State's housing cost index, with the baseline set at 100}
\item{utilities_cost_index}{State's utilities cost index, with the baseline set at 100}
\item{transportation_cost_index}{State's transportation cost index, with the baseline set at 100}
\item{misc_cost_index}{State's miscellaneous cost index, with the baseline set at 100}
\item{minimum_wage}{State's minimum wage (hourly rate)}
}
}
}
\source{
Download from
\url{https://worldpopulationreview.com/state-rankings/cost-of-living-index-by-state},
\url{https://worldpopulationreview.com/state-rankings/cost-of-living-index-by-state}, and
\url{https://www.laborlawcenter.com/state-minimum-wage-rates}.
}
\usage{
us_cost_of_living
}
\description{
A data set that contains information related to cost of living in the US.
}
\keyword{datasets}
