% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_datasets.R
\docType{data}
\name{us_state_local_tax_rate}
\alias{us_state_local_tax_rate}
\title{US state and average local tax rate data set}
\format{
\subsection{\code{us_state_local_tax_rate}}{

A data frame with 50 rows representing 50 states and 5 columns
\describe{
\item{state}{State name}
\item{state_abbr}{State name abbreviation}
\item{state_tax_rate}{State's tax rate, in percent}
\item{avg_local_tax_rate}{State's average local tax rate, in percent}
\item{combined_tax_rate}{State's combined state and local tax rate, in percent}
}
}
}
\source{
Download from
\url{https://worldpopulationreview.com/state-rankings/sales-tax-by-state}.
}
\usage{
us_state_local_tax_rate
}
\description{
A data set that contains information related to US state and average local tax rate data set.
}
\keyword{datasets}
